requirements.txt
beautifulsoup4
requests
apache-airflow
csv